requirements.txt
# Core API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.1

# Data Processing
bibtexparser>=1.4.1
pandas>=2.1.0
numpy>=1.24.0

# arXiv API
arxiv>=2.1.0

# Text Processing
python-Levenshtein>=0.21.0
rapidfuzz>=3.5.0

# HTTP/Networking
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type Checking
mypy>=1.7.0
types-requests>=2.31.0
types-PyYAML>=6.0.0
